+++ ChangeLog.md
@@ -4,7 +4,12 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/)
 and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.html).
 
-## [0.35.0]
+## [0.35.2]
+### Changed
+- Corrected a bug that could leave `sets=True` and `gets=True` in property and method
+  paramattr descriptors in cases where getter and setter (respectively) were unimplemented
+
+## [0.35.1]
 ### Added
 - Paramattr now supports a new boolean keyword argument, `inherit`, which takes 
   defaults for the other paramattr keyword arguments from the paramattr in the parent